refactor: create catalog healpers
