+++ src/helpers/getUsername.js
@@ -1,8 +1,6 @@
-const getUsername = () => {
+export const getUsername = () => {
   const args = process.argv; // массив аргументов
   const userNameArg = args.find((str) => str.startsWith("--username=")); // аргумент с именем пользователя
   const username = userNameArg ? userNameArg.slice(11) : "Anonymous"; // имя пользователя
   return username;
 };
-
-export default getUsername;
